moved stacked charts test back to non-UI suite
