+++ info.limpet.test/src/info/limpet/data2/TestStackedCharts.java
@@ -1,4 +1,4 @@
-package info.limpet.rcp;
+package info.limpet.data2;
 
 import info.limpet.IStoreItem;
 import info.limpet.impl.NumberDocument;
@@ -15,21 +15,24 @@ import org.junit.Test;
 
 public class TestStackedCharts extends TestCase
 {
- 
+
   @Test
   public void testGenChartModel()
   {
-    ArrayList<IStoreItem> selection = new ArrayList<IStoreItem>();
-    StoreGroup data = new SampleData().getData(20);
-    
-    NumberDocument product = (NumberDocument) data.get("Speed One Time + Speed Two Time + Speed two irregular time");
+    final ArrayList<IStoreItem> selection = new ArrayList<IStoreItem>();
+    final StoreGroup data = new SampleData().getData(20);
+
+    final NumberDocument product =
+        (NumberDocument) data
+            .get("Speed One Time + Speed Two Time + Speed two irregular time");
     assertEquals("17 values", 17, product.size());
     assertEquals("17 indices", 17, product.getIndexValues().getSize());
-    
+
     assertNotNull(product);
-    
+
     selection.add(product);
-    ChartSet model = ShowInStackedChartsOverview.createModelFor(selection);
+    final ChartSet model =
+        ShowInStackedChartsOverview.createModelFor(selection);
     assertNotNull(model);
   }
