+++ info.limpet.test/src/info/limpet/rcp/AllRcpTests.java
@@ -1,14 +1,13 @@
 package info.limpet.rcp;
 
+import junit.framework.TestSuite;
+
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
-import junit.framework.TestSuite;
-
 @Suite.SuiteClasses(
 { 
-	TestReflectivePropertySource.class,
-	TestStackedCharts.class
+	TestReflectivePropertySource.class
 })
 
 @RunWith(Suite.class)
